Runtime compiles

+++ runtime/src/constants.rs
@@ -67,10 +67,10 @@ pub mod time {
     // NOTE: Currently it is not possible to change the epoch duration after the chain has started.
     //       Attempting to do so will brick block production.
     pub const EPOCH_DURATION_IN_BLOCKS: BlockNumber = 10 * MINUTES;
-    pub const EPOCH_DURATION_IN_SLOTS: u64 = {
+    pub const EPOCH_DURATION_IN_SLOTS: u32 = {
         const SLOT_FILL_RATE: f64 = MILLISECS_PER_BLOCK as f64 / SLOT_DURATION as f64;
 
-        (EPOCH_DURATION_IN_BLOCKS as f64 * SLOT_FILL_RATE) as u64
+        (EPOCH_DURATION_IN_BLOCKS as f64 * SLOT_FILL_RATE) as u32
     };
 
     // These time units are defined in number of blocks.
